bedrock/deepseek/requirements.txt
langchain_aws >=0.2.12, <=0.4
langchain >=0.3.17, <=0.4
transformers >=4.48.2, <=4.49
boto3 >=1.36.11, <=1.37
python-dotenv >=1.0.1, <=1.1